Created ProviderDetectionTestingFramework
Created ProviderDetectionTestingFramework and refactored the provider detection code
