+++ CryptoAnalysis/src/test/java/tests/providerdetection/ProviderDetectionTests.java
@@ -11,7 +11,7 @@ public class ProviderDetectionTests {
 	// Checks if provider of type `java.security.Provider` is detected when given as a variable
 	@Test
 	public void providerDetectionTest1() {
-		ProviderDetection providerDetection = new ProviderDetection();
+		ProviderDetectionTestingFramework providerDetection = new ProviderDetectionTestingFramework();
 		String sootClassPath = providerDetection.getSootClassPath();
 		String mainClass = "tests.providerdetection.ProviderDetectionExample1"; 
 		providerDetection.setupSoot(sootClassPath, mainClass);
@@ -25,7 +25,7 @@ public class ProviderDetectionTests {
 	// Checks if provider of type `java.security.Provider` is detected when given directly
 	@Test
 	public void providerDetectionTest2() {
-		ProviderDetection providerDetection = new ProviderDetection();
+		ProviderDetectionTestingFramework providerDetection = new ProviderDetectionTestingFramework();
 		String sootClassPath = providerDetection.getSootClassPath();
 		String mainClass = "tests.providerdetection.ProviderDetectionExample2"; 
 		providerDetection.setupSoot(sootClassPath, mainClass);
@@ -41,7 +41,7 @@ public class ProviderDetectionTests {
 	// is given as a variable, and the rules for that provider exist
 	@Test
 	public void providerDetectionTest3() {
-		ProviderDetection providerDetection = new ProviderDetection();
+		ProviderDetectionTestingFramework providerDetection = new ProviderDetectionTestingFramework();
 		String sootClassPath = providerDetection.getSootClassPath();
 		String mainClass = "tests.providerdetection.ProviderDetectionExample1"; 
 		providerDetection.setupSoot(sootClassPath, mainClass);
@@ -56,7 +56,7 @@ public class ProviderDetectionTests {
 	// is given directly, and the rules for that provider exist
 	@Test
 	public void providerDetectionTest4() {
-		ProviderDetection providerDetection = new ProviderDetection();
+		ProviderDetectionTestingFramework providerDetection = new ProviderDetectionTestingFramework();
 		String sootClassPath = providerDetection.getSootClassPath();
 		String mainClass = "tests.providerdetection.ProviderDetectionExample2"; 
 		providerDetection.setupSoot(sootClassPath, mainClass);
@@ -70,7 +70,7 @@ public class ProviderDetectionTests {
 	// is given as a variable, and the rules for that provider exist
 	@Test
 	public void providerDetectionTest5() {
-		ProviderDetection providerDetection = new ProviderDetection();
+		ProviderDetectionTestingFramework providerDetection = new ProviderDetectionTestingFramework();
 		String sootClassPath = providerDetection.getSootClassPath();
 		String mainClass = "tests.providerdetection.ProviderDetectionExample3"; 
 		providerDetection.setupSoot(sootClassPath, mainClass);
@@ -84,7 +84,7 @@ public class ProviderDetectionTests {
 	// is given directly, and the rules for that provider exist
 	@Test
 	public void providerDetectionTest6() {
-		ProviderDetection providerDetection = new ProviderDetection();
+		ProviderDetectionTestingFramework providerDetection = new ProviderDetectionTestingFramework();
 		String sootClassPath = providerDetection.getSootClassPath();
 		String mainClass = "tests.providerdetection.ProviderDetectionExample4"; 
 		providerDetection.setupSoot(sootClassPath, mainClass);
@@ -97,7 +97,7 @@ public class ProviderDetectionTests {
 	// Checks if provider of type `java.lang.String` is detected when given as a variable
 	@Test
 	public void providerDetectionTest7() {
-		ProviderDetection providerDetection = new ProviderDetection();
+		ProviderDetectionTestingFramework providerDetection = new ProviderDetectionTestingFramework();
 		String sootClassPath = providerDetection.getSootClassPath();
 		String mainClass = "tests.providerdetection.ProviderDetectionExample5"; 
 		providerDetection.setupSoot(sootClassPath, mainClass);
@@ -111,7 +111,7 @@ public class ProviderDetectionTests {
 	// Checks if provider of type `java.lang.String` is detected when given directly
 	@Test
 	public void providerDetectionTest8() {
-		ProviderDetection providerDetection = new ProviderDetection();
+		ProviderDetectionTestingFramework providerDetection = new ProviderDetectionTestingFramework();
 		String sootClassPath = providerDetection.getSootClassPath();
 		String mainClass = "tests.providerdetection.ProviderDetectionExample6"; 
 		providerDetection.setupSoot(sootClassPath, mainClass);
@@ -126,7 +126,7 @@ public class ProviderDetectionTests {
 	// is given as a variable, and the rules for that provider exist
 	@Test
 	public void providerDetectionTest9() {
-		ProviderDetection providerDetection = new ProviderDetection();
+		ProviderDetectionTestingFramework providerDetection = new ProviderDetectionTestingFramework();
 		String sootClassPath = providerDetection.getSootClassPath();
 		String mainClass = "tests.providerdetection.ProviderDetectionExample5"; 
 		providerDetection.setupSoot(sootClassPath, mainClass);
@@ -140,7 +140,7 @@ public class ProviderDetectionTests {
 	// is given directly, and the rules for that provider exist
 	@Test
 	public void providerDetectionTest10() {
-		ProviderDetection providerDetection = new ProviderDetection();
+		ProviderDetectionTestingFramework providerDetection = new ProviderDetectionTestingFramework();
 		String sootClassPath = providerDetection.getSootClassPath();
 		String mainClass = "tests.providerdetection.ProviderDetectionExample6"; 
 		providerDetection.setupSoot(sootClassPath, mainClass);
@@ -154,7 +154,7 @@ public class ProviderDetectionTests {
 	// is given as a variable, and the rules for that provider exist
 	@Test
 	public void providerDetectionTest11() {
-		ProviderDetection providerDetection = new ProviderDetection();
+		ProviderDetectionTestingFramework providerDetection = new ProviderDetectionTestingFramework();
 		String sootClassPath = providerDetection.getSootClassPath();
 		String mainClass = "tests.providerdetection.ProviderDetectionExample7"; 
 		providerDetection.setupSoot(sootClassPath, mainClass);
@@ -168,7 +168,7 @@ public class ProviderDetectionTests {
 	// is given directly, and the rules for that provider exist
 	@Test
 	public void providerDetectionTest12() {
-		ProviderDetection providerDetection = new ProviderDetection();
+		ProviderDetectionTestingFramework providerDetection = new ProviderDetectionTestingFramework();
 		String sootClassPath = providerDetection.getSootClassPath();
 		String mainClass = "tests.providerdetection.ProviderDetectionExample8"; 
 		providerDetection.setupSoot(sootClassPath, mainClass);
@@ -182,7 +182,7 @@ public class ProviderDetectionTests {
 	// flows through TERNARY operators
 	@Test
 	public void providerDetectionTest13() {
-		ProviderDetection providerDetection = new ProviderDetection();
+		ProviderDetectionTestingFramework providerDetection = new ProviderDetectionTestingFramework();
 		String sootClassPath = providerDetection.getSootClassPath();
 		String mainClass = "tests.providerdetection.ProviderDetectionExample9"; 
 		providerDetection.setupSoot(sootClassPath, mainClass);
@@ -196,7 +196,7 @@ public class ProviderDetectionTests {
 	// flows through IF-ELSE statements
 	@Test
 	public void providerDetectionTest14() {
-		ProviderDetection providerDetection = new ProviderDetection();
+		ProviderDetectionTestingFramework providerDetection = new ProviderDetectionTestingFramework();
 		String sootClassPath = providerDetection.getSootClassPath();
 		String mainClass = "tests.providerdetection.ProviderDetectionExample10"; 
 		providerDetection.setupSoot(sootClassPath, mainClass);
@@ -210,7 +210,7 @@ public class ProviderDetectionTests {
 	// flows through SWITCH statements
 	@Test
 	public void providerDetectionTest15() {
-		ProviderDetection providerDetection = new ProviderDetection();
+		ProviderDetectionTestingFramework providerDetection = new ProviderDetectionTestingFramework();
 		String sootClassPath = providerDetection.getSootClassPath();
 		String mainClass = "tests.providerdetection.ProviderDetectionExample11"; 
 		providerDetection.setupSoot(sootClassPath, mainClass);
@@ -224,7 +224,7 @@ public class ProviderDetectionTests {
 	// flows through TERNARY operators
 	@Test
 	public void providerDetectionTest16() {
-		ProviderDetection providerDetection = new ProviderDetection();
+		ProviderDetectionTestingFramework providerDetection = new ProviderDetectionTestingFramework();
 		String sootClassPath = providerDetection.getSootClassPath();
 		String mainClass = "tests.providerdetection.ProviderDetectionExample12"; 
 		providerDetection.setupSoot(sootClassPath, mainClass);
@@ -238,7 +238,7 @@ public class ProviderDetectionTests {
 	// flows through IF-ELSE statements
 	@Test
 	public void providerDetectionTest17() {
-		ProviderDetection providerDetection = new ProviderDetection();
+		ProviderDetectionTestingFramework providerDetection = new ProviderDetectionTestingFramework();
 		String sootClassPath = providerDetection.getSootClassPath();
 		String mainClass = "tests.providerdetection.ProviderDetectionExample13"; 
 		providerDetection.setupSoot(sootClassPath, mainClass);
@@ -252,7 +252,7 @@ public class ProviderDetectionTests {
 	// flows through SWITCH statements
 	@Test
 	public void providerDetectionTest18() {
-		ProviderDetection providerDetection = new ProviderDetection();
+		ProviderDetectionTestingFramework providerDetection = new ProviderDetectionTestingFramework();
 		String sootClassPath = providerDetection.getSootClassPath();
 		String mainClass = "tests.providerdetection.ProviderDetectionExample14"; 
 		providerDetection.setupSoot(sootClassPath, mainClass);
